docs: port uxdot-search to typescript
