+++ .gitignore
@@ -14,6 +14,7 @@ docs/assets/javascript/elements/uxdot-header.js
 docs/assets/javascript/elements/uxdot-masthead.js
 docs/assets/javascript/elements/uxdot-feedback.js
 docs/assets/javascript/elements/uxdot-hero.js
+docs/assets/javascript/elements/uxdot-search.js
 docs/assets/javascript/elements/uxdot-copy-button.js
 docs/assets/javascript/elements/uxdot-copy-permalink.js
 docs/_data/playgrounds.json
