+++ docs/assets/javascript/elements/uxdot-search.css
@@ -1,0 +1,67 @@
+:host {
+  display: grid;
+  gap: var(--rh-space-lg);
+  grid-template: 1fr / 1fr max-content;
+  font-family: var(--rh-font-family-body-text);
+  position: relative;
+}
+
+[hidden] { display: none !important; }
+
+input {
+  border: var(--rh-border-width-sm);
+  border-bottom-color: var(--rh-color-gray-50);
+  padding: var(--rh-space-md);
+}
+
+input:focus {
+  border-bottom-color: var(--rh-color-interactive-blue-darker);
+  border-bottom-width: var(--rh-border-width-md);
+}
+
+input::placeholder {
+  font-family: inherit;
+  font-size: var(--rh-font-size-body-text-md);
+}
+
+#container {
+  position: absolute;
+  max-height: 300px;
+  overflow-y: scroll;
+  z-index: 2;
+  grid-column: 1/2;
+  width: calc(100% + 2 * var(--rh-border-width-md));
+  inset-block-start: var(--rh-length-2xl);
+  inset-inline-start: calc(-1 * var(--rh-space-md));
+  padding: var(--rh-space-sm);
+}
+
+ol {
+  list-style-type: none;
+  flex-flow: column nowrap;
+  padding-inline-start: 0;
+  border: var(--rh-border-width-sm);
+  background: var(--rh-color-surface-lightest);
+  margin: 0;
+  height: calc(100% - var(--rh-space-md));
+  width: calc(100% - 2 * var(--rh-border-width-md));
+}
+
+a {
+  color: inherit;
+  text-decoration: none;
+  font-size: var(--rh-font-size-body-text-sm);
+}
+
+a:focus {
+  outline: none;
+}
+
+li {
+  padding: var(--rh-space-md);
+}
+
+li[aria-selected="true"],
+li:focus-within {
+  outline: var(--rh-border-width-md);
+}
